New SOP encoding to AIG
This encoding has four layers:
A. product of UAP
B. sum of product of UAP || sum of one-hot states
C. product of A and B
D. sum of C for every CAP/latch

The idea is that this might give ABC more opportunities to optimize, as "likely related" gates are closer together. The initial encoding is also smaller than the other methods in many cases.